Clone elements in articleContentToHTML.serialize
Since we depend on the `node.type` of some children (IE: children of
tables, like captions), we cannot change the type by wrapping them in
fragments to set their keys. Therefore we need to use
`React.cloneElement` instead.

A slightly more performant version of this could probably be to set the
key internally in every serializer, but that would probably be somewhat
complex.