fix(auth): token refresh race condition after expiration
This patch fixes a race condition where we called `renewAuth` multiple
times, leading to us calling the refresh endpoint with the same
refresh token multiple times. Since a refresh token is only supposed to
be used once, this lead to the refresh endpoint clearing the cookie
since it got an error on refresh.